refacor: use sub command
